editor page bundle sizes were huge again; reduced by moving graphql stuff
